split i18n into email + ui mods
