+++ src/models/src/email.rs
@@ -1,12 +1,12 @@
 use crate::app_state::AppState;
 use crate::entity::magic_links::MagicLink;
 use crate::entity::users::User;
-use crate::i18n::email_change_info_new::I18nEmailChangeInfoNew;
-use crate::i18n::email_confirm_change::I18nEmailConfirmChange;
-use crate::i18n::email_password_new::I18nEmailPasswordNew;
-use crate::i18n::email_reset::I18nEmailReset;
-use crate::i18n::email_reset_info::I18nEmailResetInfo;
-use crate::i18n::SsrJson;
+use crate::i18n::email::change_info_new::I18nEmailChangeInfoNew;
+use crate::i18n::email::confirm_change::I18nEmailConfirmChange;
+use crate::i18n::email::password_new::I18nEmailPasswordNew;
+use crate::i18n::email::reset::I18nEmailReset;
+use crate::i18n::email::reset_info::I18nEmailResetInfo;
+use crate::i18n::ui::SsrJson;
 use actix_web::web;
 use askama_actix::Template;
 use chrono::DateTime;
