+++ src/models/src/i18n/email/change_info_old.rs
@@ -1,4 +1,4 @@
-use crate::i18n::SsrJson;
+use crate::i18n::ui::SsrJson;
 use crate::language::Language;
 use serde::Serialize;
