+++ src/models/src/i18n/email/mod.rs
@@ -1,0 +1,7 @@
+pub mod change_info_new;
+pub mod change_info_old;
+pub mod confirm_change;
+pub mod confirm_change_html;
+pub mod password_new;
+pub mod reset;
+pub mod reset_info;
