+++ src/models/src/i18n/email/password_new.rs
@@ -1,4 +1,4 @@
-use crate::i18n::SsrJson;
+use crate::i18n::ui::SsrJson;
 use crate::language::Language;
 use once_cell::sync::Lazy;
 use serde::Serialize;
