+++ src/models/src/i18n/ui/account.rs
@@ -1,5 +1,5 @@
-use crate::i18n::password_policy::I18nPasswordPolicy;
-use crate::i18n::SsrJson;
+use crate::i18n::ui::password_policy::I18nPasswordPolicy;
+use crate::i18n::ui::SsrJson;
 use crate::language::Language;
 use serde::Serialize;
