+++ src/models/src/i18n/ui/register.rs
@@ -1,4 +1,4 @@
-use crate::i18n::SsrJson;
+use crate::i18n::ui::SsrJson;
 use crate::language::Language;
 use serde::Serialize;
