+++ src/models/src/templates.rs
@@ -1,15 +1,15 @@
 use crate::entity::colors::Colors;
 use crate::entity::password::PasswordPolicy;
-use crate::i18n::account::I18nAccount;
-use crate::i18n::authorize::I18nAuthorize;
-use crate::i18n::device::I18nDevice;
-use crate::i18n::email_confirm_change_html::I18nEmailConfirmChangeHtml;
-use crate::i18n::error::I18nError;
-use crate::i18n::index::I18nIndex;
-use crate::i18n::logout::I18nLogout;
-use crate::i18n::password_reset::I18nPasswordReset;
-use crate::i18n::register::I18nRegister;
-use crate::i18n::SsrJson;
+use crate::i18n::email::confirm_change_html::I18nEmailConfirmChangeHtml;
+use crate::i18n::ui::account::I18nAccount;
+use crate::i18n::ui::authorize::I18nAuthorize;
+use crate::i18n::ui::device::I18nDevice;
+use crate::i18n::ui::error::I18nError;
+use crate::i18n::ui::index::I18nIndex;
+use crate::i18n::ui::logout::I18nLogout;
+use crate::i18n::ui::password_reset::I18nPasswordReset;
+use crate::i18n::ui::register::I18nRegister;
+use crate::i18n::ui::SsrJson;
 use crate::language::Language;
 use actix_web::http::StatusCode;
 use actix_web::{HttpResponse, HttpResponseBuilder};
